Moving stat info over.
