+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/components/Divider.kt
@@ -1,0 +1,54 @@
+package com.adammcneilly.pocketleague.composables.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+/**
+ * <a href="https://material.io/components/dividers" class="external" target="_blank">Material Design divider</a>.
+ *
+ * A divider is a thin line that groups content in lists and layouts.
+ *
+ * ![Divider image](https://developer.android.com/images/reference/androidx/compose/material3/divider.png)
+ *
+ * We needed to copy this code over since a divider does not exist in compose multiplatform yet.
+ *
+ * @param modifier the [Modifier] to be applied to this divider line
+ * @param color color of this divider line
+ * @param thickness thickness of this divider line. Using [Dp.Hairline] will produce a single pixel
+ * divider regardless of screen density.
+ * @param startIndent start offset of this line. No offset by default.
+ */
+@Composable
+fun Divider(
+    modifier: Modifier = Modifier,
+    color: Color = MaterialTheme.colorScheme.onSurfaceVariant,
+    thickness: Dp = 1.dp,
+    startIndent: Dp = 0.dp
+) {
+    val indentMod = if (startIndent.value != 0f) {
+        Modifier.padding(start = startIndent)
+    } else {
+        Modifier
+    }
+    val targetThickness = if (thickness == Dp.Hairline) {
+        (1f / LocalDensity.current.density).dp
+    } else {
+        thickness
+    }
+    Box(
+        modifier.then(indentMod)
+            .fillMaxWidth()
+            .height(targetThickness)
+            .background(color = color)
+    )
+}
